updated app bar
